refactor: css 코드 및 util 함수 분리
